Add TemplateFrame to templates
