+++ docs/data/material/getting-started/templates/blog/Blog.js
@@ -6,7 +6,7 @@ import AppAppBar from './components/AppAppBar';
 import MainContent from './components/MainContent';
 import Latest from './components/Latest';
 import Footer from './components/Footer';
-import NavBar from './NavBar';
+import TemplateFrame from './TemplateFrame';
 
 import getBlogTheme from './theme/getBlogTheme';
 
@@ -41,25 +41,26 @@ export default function Blog() {
   };
 
   return (
-    <ThemeProvider theme={showCustomTheme ? blogTheme : defaultTheme}>
-      <CssBaseline />
-      {/* you can delete this NavBar component since it's just no navigate to other pages */}
-      <NavBar
-        toggleCustomTheme={toggleCustomTheme}
-        showCustomTheme={showCustomTheme}
-        mode={mode}
-        toggleColorMode={toggleColorMode}
-      />
-      <AppAppBar />
-      <Container
-        maxWidth="lg"
-        component="main"
-        sx={{ display: 'flex', flexDirection: 'column', mt: 24, mb: 16, gap: 4 }}
-      >
-        <MainContent />
-        <Latest />
-      </Container>
-      <Footer />
-    </ThemeProvider>
+    // you can delete this TemplateFrame component since it's just no navigate to other pages
+    <TemplateFrame
+      toggleCustomTheme={toggleCustomTheme}
+      showCustomTheme={showCustomTheme}
+      mode={mode}
+      toggleColorMode={toggleColorMode}
+    >
+      <ThemeProvider theme={showCustomTheme ? blogTheme : defaultTheme}>
+        <CssBaseline />
+        <AppAppBar />
+        <Container
+          maxWidth="lg"
+          component="main"
+          sx={{ display: 'flex', flexDirection: 'column', my: 16, gap: 4 }}
+        >
+          <MainContent />
+          <Latest />
+        </Container>
+        <Footer />
+      </ThemeProvider>
+    </TemplateFrame>
   );
 }
