+++ docs/data/material/getting-started/templates/checkout/Checkout.tsx
@@ -20,7 +20,7 @@ import InfoMobile from './components/InfoMobile';
 import PaymentForm from './components/PaymentForm';
 import Review from './components/Review';
 import SitemarkIcon from './components/SitemarkIcon';
-import NavBar from './NavBar';
+import TemplateFrame from './TemplateFrame';
 
 const steps = ['Shipping address', 'Payment details', 'Review your order'];
 function getStepContent(step: number) {
@@ -71,216 +71,218 @@ export default function Checkout() {
     setActiveStep(activeStep - 1);
   };
   return (
-    <ThemeProvider theme={showCustomTheme ? checkoutTheme : defaultTheme}>
-      <CssBaseline />
-      {/* you can delete this NavBar component since it's just no navigate to other pages */}
-      <NavBar
-        toggleCustomTheme={toggleCustomTheme}
-        showCustomTheme={showCustomTheme}
-        mode={mode}
-        toggleColorMode={toggleColorMode}
-      />
-      <Grid container sx={{ height: { xs: '100%', sm: '100dvh' } }}>
-        <Grid
-          size={{ xs: 12, sm: 5, lg: 4 }}
-          sx={{
-            display: { xs: 'none', md: 'flex' },
-            flexDirection: 'column',
-            backgroundColor: 'background.paper',
-            borderRight: { sm: 'none', md: '1px solid' },
-            borderColor: { sm: 'none', md: 'divider' },
-            alignItems: 'start',
-            pt: 24,
-            px: 10,
-            gap: 4,
-          }}
-        >
-          <SitemarkIcon />
-          <Box
+    // You can delete this NavBar component since it's just no navigate to other pages
+    <TemplateFrame
+      toggleCustomTheme={toggleCustomTheme}
+      showCustomTheme={showCustomTheme}
+      mode={mode}
+      toggleColorMode={toggleColorMode}
+    >
+      <ThemeProvider theme={showCustomTheme ? checkoutTheme : defaultTheme}>
+        <CssBaseline />
+
+        <Grid container sx={{ height: { xs: '100%', sm: '100dvh' } }}>
+          <Grid
+            size={{ xs: 12, sm: 5, lg: 4 }}
+            sx={{
+              display: { xs: 'none', md: 'flex' },
+              flexDirection: 'column',
+              backgroundColor: 'background.paper',
+              borderRight: { sm: 'none', md: '1px solid' },
+              borderColor: { sm: 'none', md: 'divider' },
+              alignItems: 'start',
+              pt: 16,
+              px: 10,
+              gap: 4,
+            }}
+          >
+            <SitemarkIcon />
+            <Box
+              sx={{
+                display: 'flex',
+                flexDirection: 'column',
+                flexGrow: 1,
+                width: '100%',
+                maxWidth: 500,
+              }}
+            >
+              <Info totalPrice={activeStep >= 2 ? '$144.97' : '$134.98'} />
+            </Box>
+          </Grid>
+          <Grid
+            size={{ sm: 12, md: 7, lg: 8 }}
             sx={{
               display: 'flex',
               flexDirection: 'column',
-              flexGrow: 1,
+              maxWidth: '100%',
               width: '100%',
-              maxWidth: 500,
-            }}
-          >
-            <Info totalPrice={activeStep >= 2 ? '$144.97' : '$134.98'} />
-          </Box>
-        </Grid>
-        <Grid
-          size={{ sm: 12, md: 7, lg: 8 }}
-          sx={{
-            display: 'flex',
-            flexDirection: 'column',
-            maxWidth: '100%',
-            width: '100%',
-            backgroundColor: { xs: 'transparent', sm: 'background.default' },
-            alignItems: 'start',
-            pt: { xs: 6, sm: 24 },
-            px: { xs: 2, sm: 10 },
-            gap: { xs: 4, md: 8 },
-          }}
-        >
-          <Box
-            sx={{
-              display: 'flex',
-              justifyContent: { sm: 'space-between', md: 'flex-end' },
-              alignItems: 'center',
-              width: '100%',
-              maxWidth: { sm: '100%', md: 600 },
+              backgroundColor: { xs: 'transparent', sm: 'background.default' },
+              alignItems: 'start',
+              pt: { xs: 6, sm: 16 },
+              px: { xs: 2, sm: 10 },
+              gap: { xs: 4, md: 8 },
             }}
           >
             <Box
               sx={{
-                display: { xs: 'none', md: 'flex' },
+                display: 'flex',
+                justifyContent: { sm: 'space-between', md: 'flex-end' },
+                alignItems: 'center',
+                width: '100%',
+                maxWidth: { sm: '100%', md: 600 },
+              }}
+            >
+              <Box
+                sx={{
+                  display: { xs: 'none', md: 'flex' },
+                  flexDirection: 'column',
+                  justifyContent: 'space-between',
+                  alignItems: 'flex-end',
+                  flexGrow: 1,
+                }}
+              >
+                <Stepper
+                  id="desktop-stepper"
+                  activeStep={activeStep}
+                  sx={{ width: '100%', height: 40 }}
+                >
+                  {steps.map((label) => (
+                    <Step
+                      sx={{ ':first-child': { pl: 0 }, ':last-child': { pr: 0 } }}
+                      key={label}
+                    >
+                      <StepLabel>{label}</StepLabel>
+                    </Step>
+                  ))}
+                </Stepper>
+              </Box>
+            </Box>
+            <Card sx={{ display: { xs: 'flex', md: 'none' }, width: '100%' }}>
+              <CardContent
+                sx={{
+                  display: 'flex',
+                  width: '100%',
+                  alignItems: 'center',
+                  justifyContent: 'space-between',
+                }}
+              >
+                <div>
+                  <Typography variant="subtitle2" gutterBottom>
+                    Selected products
+                  </Typography>
+                  <Typography variant="body1">
+                    {activeStep >= 2 ? '$144.97' : '$134.98'}
+                  </Typography>
+                </div>
+                <InfoMobile totalPrice={activeStep >= 2 ? '$144.97' : '$134.98'} />
+              </CardContent>
+            </Card>
+            <Box
+              sx={{
+                display: 'flex',
                 flexDirection: 'column',
-                justifyContent: 'space-between',
-                alignItems: 'flex-end',
                 flexGrow: 1,
+                width: '100%',
+                maxWidth: { sm: '100%', md: 600 },
+                maxHeight: '720px',
+                gap: { xs: 5, md: 'none' },
               }}
             >
               <Stepper
-                id="desktop-stepper"
+                id="mobile-stepper"
                 activeStep={activeStep}
-                sx={{ width: '100%', height: 40 }}
+                alternativeLabel
+                sx={{ display: { sm: 'flex', md: 'none' } }}
               >
                 {steps.map((label) => (
                   <Step
-                    sx={{ ':first-child': { pl: 0 }, ':last-child': { pr: 0 } }}
+                    sx={{
+                      ':first-child': { pl: 0 },
+                      ':last-child': { pr: 0 },
+                      '& .MuiStepConnector-root': { top: { xs: 6, sm: 12 } },
+                    }}
                     key={label}
                   >
-                    <StepLabel>{label}</StepLabel>
+                    <StepLabel
+                      sx={{ '.MuiStepLabel-labelContainer': { maxWidth: '70px' } }}
+                    >
+                      {label}
+                    </StepLabel>
                   </Step>
                 ))}
               </Stepper>
-            </Box>
-          </Box>
-          <Card sx={{ display: { xs: 'flex', md: 'none' }, width: '100%' }}>
-            <CardContent
-              sx={{
-                display: 'flex',
-                width: '100%',
-                alignItems: 'center',
-                justifyContent: 'space-between',
-              }}
-            >
-              <div>
-                <Typography variant="subtitle2" gutterBottom>
-                  Selected products
-                </Typography>
-                <Typography variant="body1">
-                  {activeStep >= 2 ? '$144.97' : '$134.98'}
-                </Typography>
-              </div>
-              <InfoMobile totalPrice={activeStep >= 2 ? '$144.97' : '$134.98'} />
-            </CardContent>
-          </Card>
-          <Box
-            sx={{
-              display: 'flex',
-              flexDirection: 'column',
-              flexGrow: 1,
-              width: '100%',
-              maxWidth: { sm: '100%', md: 600 },
-              maxHeight: '720px',
-              gap: { xs: 5, md: 'none' },
-            }}
-          >
-            <Stepper
-              id="mobile-stepper"
-              activeStep={activeStep}
-              alternativeLabel
-              sx={{ display: { sm: 'flex', md: 'none' } }}
-            >
-              {steps.map((label) => (
-                <Step
-                  sx={{
-                    ':first-child': { pl: 0 },
-                    ':last-child': { pr: 0 },
-                    '& .MuiStepConnector-root': { top: { xs: 6, sm: 12 } },
-                  }}
-                  key={label}
-                >
-                  <StepLabel
-                    sx={{ '.MuiStepLabel-labelContainer': { maxWidth: '70px' } }}
-                  >
-                    {label}
-                  </StepLabel>
-                </Step>
-              ))}
-            </Stepper>
-            {activeStep === steps.length ? (
-              <Stack spacing={2} useFlexGap>
-                <Typography variant="h1">📦</Typography>
-                <Typography variant="h5">Thank you for your order!</Typography>
-                <Typography variant="body1" sx={{ color: 'text.secondary' }}>
-                  Your order number is
-                  <strong>&nbsp;#140396</strong>. We have emailed your order
-                  confirmation and will update you once its shipped.
-                </Typography>
-                <Button
-                  variant="contained"
-                  sx={{ alignSelf: 'start', width: { xs: '100%', sm: 'auto' } }}
-                >
-                  Go to my orders
-                </Button>
-              </Stack>
-            ) : (
-              <React.Fragment>
-                {getStepContent(activeStep)}
-                <Box
-                  sx={[
-                    {
-                      display: 'flex',
-                      flexDirection: { xs: 'column-reverse', sm: 'row' },
-                      alignItems: 'end',
-                      flexGrow: 1,
-                      gap: 1,
-                      pb: { xs: 12, sm: 0 },
-                      mt: { xs: 2, sm: 0 },
-                      mb: '60px',
-                    },
-                    activeStep !== 0
-                      ? { justifyContent: 'space-between' }
-                      : { justifyContent: 'flex-end' },
-                  ]}
-                >
-                  {activeStep !== 0 && (
-                    <Button
-                      startIcon={<ChevronLeftRoundedIcon />}
-                      onClick={handleBack}
-                      variant="text"
-                      sx={{ display: { xs: 'none', sm: 'flex' } }}
-                    >
-                      Previous
-                    </Button>
-                  )}
-                  {activeStep !== 0 && (
-                    <Button
-                      startIcon={<ChevronLeftRoundedIcon />}
-                      onClick={handleBack}
-                      variant="outlined"
-                      fullWidth
-                      sx={{ display: { xs: 'flex', sm: 'none' } }}
-                    >
-                      Previous
-                    </Button>
-                  )}
+              {activeStep === steps.length ? (
+                <Stack spacing={2} useFlexGap>
+                  <Typography variant="h1">📦</Typography>
+                  <Typography variant="h5">Thank you for your order!</Typography>
+                  <Typography variant="body1" sx={{ color: 'text.secondary' }}>
+                    Your order number is
+                    <strong>&nbsp;#140396</strong>. We have emailed your order
+                    confirmation and will update you once its shipped.
+                  </Typography>
                   <Button
                     variant="contained"
-                    endIcon={<ChevronRightRoundedIcon />}
-                    onClick={handleNext}
-                    sx={{ width: { xs: '100%', sm: 'fit-content' } }}
+                    sx={{ alignSelf: 'start', width: { xs: '100%', sm: 'auto' } }}
                   >
-                    {activeStep === steps.length - 1 ? 'Place order' : 'Next'}
+                    Go to my orders
                   </Button>
-                </Box>
-              </React.Fragment>
-            )}
-          </Box>
+                </Stack>
+              ) : (
+                <React.Fragment>
+                  {getStepContent(activeStep)}
+                  <Box
+                    sx={[
+                      {
+                        display: 'flex',
+                        flexDirection: { xs: 'column-reverse', sm: 'row' },
+                        alignItems: 'end',
+                        flexGrow: 1,
+                        gap: 1,
+                        pb: { xs: 12, sm: 0 },
+                        mt: { xs: 2, sm: 0 },
+                        mb: '60px',
+                      },
+                      activeStep !== 0
+                        ? { justifyContent: 'space-between' }
+                        : { justifyContent: 'flex-end' },
+                    ]}
+                  >
+                    {activeStep !== 0 && (
+                      <Button
+                        startIcon={<ChevronLeftRoundedIcon />}
+                        onClick={handleBack}
+                        variant="text"
+                        sx={{ display: { xs: 'none', sm: 'flex' } }}
+                      >
+                        Previous
+                      </Button>
+                    )}
+                    {activeStep !== 0 && (
+                      <Button
+                        startIcon={<ChevronLeftRoundedIcon />}
+                        onClick={handleBack}
+                        variant="outlined"
+                        fullWidth
+                        sx={{ display: { xs: 'flex', sm: 'none' } }}
+                      >
+                        Previous
+                      </Button>
+                    )}
+                    <Button
+                      variant="contained"
+                      endIcon={<ChevronRightRoundedIcon />}
+                      onClick={handleNext}
+                      sx={{ width: { xs: '100%', sm: 'fit-content' } }}
+                    >
+                      {activeStep === steps.length - 1 ? 'Place order' : 'Next'}
+                    </Button>
+                  </Box>
+                </React.Fragment>
+              )}
+            </Box>
+          </Grid>
         </Grid>
-      </Grid>
-    </ThemeProvider>
+      </ThemeProvider>
+    </TemplateFrame>
   );
 }
