+++ docs/data/material/getting-started/templates/dashboard/Dashboard.js
@@ -8,7 +8,7 @@ import AppNavbar from './components/AppNavbar';
 import Header from './components/Header';
 import MainGrid from './components/MainGrid';
 import SideMenu from './components/SideMenu';
-import NavBar from './NavBar';
+import TemplateFrame from './TemplateFrame';
 
 export default function Dashboard() {
   const [mode, setMode] = React.useState('light');
@@ -41,44 +41,42 @@ export default function Dashboard() {
   };
 
   return (
-    <ThemeProvider theme={showCustomTheme ? dashboardTheme : defaultTheme}>
-      <CssBaseline />
-      {/* you can delete this NavBar component since it's just no navigate to other pages */}
-      <NavBar
-        toggleCustomTheme={toggleCustomTheme}
-        showCustomTheme={showCustomTheme}
-        mode={mode}
-        toggleColorMode={toggleColorMode}
-      />
-      <Box sx={{ display: 'flex' }}>
-        <SideMenu />
-        <AppNavbar />
-        {/* Main content */}
-        <Box
-          component="main"
-          sx={(theme) => ({
-            position: { sm: 'relative', md: '' },
-            top: { sm: '48px', md: '60px' },
-            height: { sm: 'calc(100vh - 48px)', md: '100vh' },
-            flexGrow: 1,
-            backgroundColor: alpha(theme.palette.background.default, 1),
-            overflow: 'auto',
-          })}
-        >
-          <Stack
-            spacing={2}
-            sx={{
-              alignItems: 'center',
-              mx: 3,
-              pb: 10,
-              mt: { xs: 16, sm: 10, md: 0 },
-            }}
+    // You can delete this TemplateFrame component since it's just no navigate to other pages
+    <TemplateFrame
+      toggleCustomTheme={toggleCustomTheme}
+      showCustomTheme={showCustomTheme}
+      mode={mode}
+      toggleColorMode={toggleColorMode}
+    >
+      <ThemeProvider theme={showCustomTheme ? dashboardTheme : defaultTheme}>
+        <CssBaseline />
+        <Box sx={{ display: 'flex' }}>
+          <SideMenu />
+          <AppNavbar />
+          {/* Main content */}
+          <Box
+            component="main"
+            sx={(theme) => ({
+              flexGrow: 1,
+              backgroundColor: alpha(theme.palette.background.default, 1),
+              overflow: 'auto',
+            })}
           >
-            <Header />
-            <MainGrid />
-          </Stack>
+            <Stack
+              spacing={2}
+              sx={{
+                alignItems: 'center',
+                mx: 3,
+                pb: 10,
+                mt: { xs: 8, md: 0 },
+              }}
+            >
+              <Header />
+              <MainGrid />
+            </Stack>
+          </Box>
         </Box>
-      </Box>
-    </ThemeProvider>
+      </ThemeProvider>
+    </TemplateFrame>
   );
 }
