+++ docs/data/material/getting-started/templates/marketing-page/components/Hero.tsx
@@ -43,7 +43,7 @@ export default function Hero() {
       sx={(theme) => ({
         width: '100%',
         backgroundRepeat: 'no-repeat',
-        mt: 6,
+
         backgroundImage:
           'radial-gradient(ellipse 80% 50% at 50% -20%, hsl(210, 100%, 90%), transparent)',
         ...theme.applyStyles('dark', {
