+++ docs/data/material/getting-started/templates/sign-in/SignIn.js
@@ -16,7 +16,7 @@ import { ThemeProvider, createTheme, styled } from '@mui/material/styles';
 import ForgotPassword from './ForgotPassword';
 import getSignInTheme from './theme/getSignInTheme';
 import { GoogleIcon, FacebookIcon, SitemarkIcon } from './CustomIcons';
-import NavBar from './NavBar';
+import TemplateFrame from './TemplateFrame';
 
 const Card = styled(MuiCard)(({ theme }) => ({
   display: 'flex',
@@ -25,8 +25,9 @@ const Card = styled(MuiCard)(({ theme }) => ({
   width: '100%',
   padding: theme.spacing(4),
   gap: theme.spacing(2),
+  margin: 'auto',
   [theme.breakpoints.up('sm')]: {
-    width: '450px',
+    maxWidth: '450px',
   },
   boxShadow:
     'hsla(220, 30%, 5%, 0.05) 0px 5px 15px 0px, hsla(220, 25%, 10%, 0.05) 0px 15px 35px -5px',
@@ -37,13 +38,11 @@ const Card = styled(MuiCard)(({ theme }) => ({
 }));
 
 const SignInContainer = styled(Stack)(({ theme }) => ({
-  height: 'auto',
+  height: '100%',
+  padding: 20,
   backgroundImage:
     'radial-gradient(ellipse at 50% 50%, hsl(210, 100%, 97%), hsl(0, 0%, 100%))',
   backgroundRepeat: 'no-repeat',
-  [theme.breakpoints.up('sm')]: {
-    height: '100dvh',
-  },
   ...theme.applyStyles('dark', {
     backgroundImage:
       'radial-gradient(at 50% 50%, hsla(210, 100%, 16%, 0.5), hsl(220, 30%, 5%))',
@@ -131,23 +130,16 @@ export default function SignIn() {
   };
 
   return (
-    <ThemeProvider theme={showCustomTheme ? SignInTheme : defaultTheme}>
-      <CssBaseline />
-      {/* you can delete this NavBar component since it's just no navigate to other pages */}
-      <NavBar
-        toggleCustomTheme={toggleCustomTheme}
-        showCustomTheme={showCustomTheme}
-        mode={mode}
-        toggleColorMode={toggleColorMode}
-      />
-      <SignInContainer direction="column" justifyContent="space-between">
-        <Stack
-          sx={{
-            justifyContent: 'center',
-            height: '100dvh',
-            p: 2,
-          }}
-        >
+    // You can delete this NavBar component since it's just no navigate to other pages
+    <TemplateFrame
+      toggleCustomTheme={toggleCustomTheme}
+      showCustomTheme={showCustomTheme}
+      mode={mode}
+      toggleColorMode={toggleColorMode}
+    >
+      <ThemeProvider theme={showCustomTheme ? SignInTheme : defaultTheme}>
+        <CssBaseline />
+        <SignInContainer direction="column" justifyContent="space-between">
           <Card variant="outlined">
             <SitemarkIcon />
             <Typography
@@ -261,8 +253,8 @@ export default function SignIn() {
               </Button>
             </Box>
           </Card>
-        </Stack>
-      </SignInContainer>
-    </ThemeProvider>
+        </SignInContainer>
+      </ThemeProvider>
+    </TemplateFrame>
   );
 }
